update: small refactor. (#202)
